### 2025-08-05 #8756 @editorialbot
[PRE REVIEW]: FleCSI: Flexible Computational Science Infrastructure
**Submitting author:** @JulienLoiseau (<a href="http://orcid.org/0000-0002-2116-2493">Julien Loiseau</a>)
**Repository:** https://github.com/flecsi/flecsi
**Branch with paper.md** (empty if default branch): joss2025
**Version:** v2.4.0
