### 2025-08-07 5131c3f
Fix ORCID, fix github workflow and update MPI reference
