### 2025-08-12 98aa995
Add documentation for execution and kernels
