+++ doc/sphinx/src/user-guide/execution.rst
@@ -1,0 +1,177 @@
+Execution Model
+***************
+
+This section describes the FleCSI execution model, which is based on the concepts of tasks and kernels.
+These abstractions allow users to express parallelism at different levels, from task scheduling to fine-grained data access.
+
+Schedulers
+++++++++++
+
+A FleCSI *scheduler* is an object that manages the launch of tasks.
+Where and when to execute the task is determined based on the task's parameter.
+They select the appropriate execution space based on the task and platform.
+
+Schedulers are obtained from a *control_point* when writing an action inside the control model:
+
+.. code-block:: c++
+
+   void action(control_point &cp) {
+     flecsi::scheduler s = cp.scheduler();
+     s.execute<task>(...);
+   }
+
+Tasks
++++++
+
+A task in FleCSI serves as the bridge between two core parts of the execution model.
+On the caller side, each ``process``, also referred to as a ``rank`` in the MPI terminology, executes its own copy of the ``main`` function and the associated control model.
+
+Each process executes its own instance of ``main`` and the control model, and so all processes collectively launch and schedule a task based on its :ref:`field privileges <field-accessors>`.
+The function for a task then executes, typically several times concurrently.
+Each of these executions is referred to as a *point task*; their number does not necessarily match the number of processes, and they can run on any process.
+This allows FleCSI to support flexible and scalable task-parallel execution.
+
+.. figure:: tikz/flecsi_execution_model.png
+
+   The different levels of abstraction in FleCSI.
+
+Callee Side
+^^^^^^^^^^^
+
+A task is a function that can accept certain special parameter types.
+For instance, the task ``foo`` may be defined as:
+
+.. code-block:: c++
+
+   void foo(flecsi::exec::cpu e,
+     flecsi::field<double>::accessor<ro, ro> f1,
+     flecsi::field<std::size_t>::accessor<rw, ro> f2) noexcept { /* ... */ }
+
+Tasks must be declared ``noexcept`` because they are scheduled and executed indirectly and cannot throw exceptions.
+The first parameter, ``e``, specifies the execution space where the task runs.
+Available options include `cpu`, `omp`, `gpu`, or `accelerator`.
+The `accelerator` space is flexible and will use the backend selected when compiling Kokkos (e.g., `Serial`, `OpenMP`, `CUDA`, etc.).
+This allows users to target different hardware execution spaces without modifying the task code.
+If the execution space is not explicitly specified in the callee declaration, the task will default to CPU execution.
+
+The field references provided during the task launch are automatically converted into :ref:`field-accessors`.
+These accessor types grant access to the real data, potentially after hidden memory copies or device transfers, depending on the privileges set by the user.
+
+Tasks can also handle collections of fields, using either ``std::vector`` or ``std::tuple``.
+For example, if all fields share the same type, a ``std::vector`` can be used to group them together:
+
+.. code-block:: c++
+
+   void bar(flecsi::exec::cpu e,
+     std::vector<flecsi::field<double>::accessor<ro, ro>> fv1) noexcept { /* ... */ }
+
+The corresponding invocation might look like:
+
+.. code-block:: c++
+
+   s.execute<bar>(flecsi::exec::on, std::vector{fr1, fr2, fr3});
+
+Caller Side
+^^^^^^^^^^^
+
+Tasks are launched by ``scheduler::execute``.
+This function launches tasks into the execution space specified by the user.
+For example, consider a task ``baz``:
+
+.. code-block:: c++
+
+   s.execute<baz>(flecsi::exec::on, fr1, fr2);
+
+Here, the task ``baz`` is specified as a template parameter.
+The first argument, ``flecsi::exec::on``, represents the execution space (when one is specified by the task).
[…]
