### 2025-08-12 4d0fa85
Add tikz picture and CI support
Co-authored-by: Richard Felix Berger <rberger@lanl.gov>
