+++ doc/sphinx/src/user-guide/tikz/flecsi_execution_model.tex
@@ -1,0 +1,218 @@
+\documentclass[tikz,border=5mm]{standalone}
+
+\usepackage{tikz}
+\usepackage[utf8]{inputenc}
+\usepackage[absolute,overlay]{textpos}
+
+\definecolor{comment}{gray}{0.4}
+\definecolor{keyword}{rgb}{0.6,0,0}
+\definecolor{identifier}{rgb}{0.4,0.35,0}
+\definecolor{string}{rgb}{0,0.3,0.4}
+
+\usetikzlibrary{
+  backgrounds,
+  arrows,
+  fit,
+  positioning,
+  calc,
+  matrix,
+  decorations.pathreplacing,
+  arrows.meta,
+  shapes.geometric
+}
+
+\pgfdeclarelayer{foreground}
+\pgfdeclarelayer{backbackground}
+\pgfsetlayers{backbackground,background,main,foreground}
+
+\begin{document}
+
+\def\PicFillColorMain{blue!10}
+\def\PicFillColorControl{blue!20}
+\def\PicFillColorAction{blue!30}
+\def\PicFillColorTask{blue!40}
+\def\PicFillColorPT{blue!50}
+\def\PicFillColorKernel{blue!60}
+\def\PicFillColorArrow{black!70}
+
+\definecolor{SoftOrange}{HTML}{E5B575}   % warm sand
+\definecolor{SoftPurple}{HTML}{B79BC4}   % pale lavender
+\definecolor{SoftGreen} {HTML}{8DCB90}   % muted mint
+
+\def\PicFillColorLayerA{SoftOrange}
+\def\PicFillColorLayerB{SoftGreen}
+
+\tikzset{
+  main/.style = {rectangle, draw,minimum width=2cm,fill=\PicFillColorMain},
+  control/.style = {ellipse, draw, fill=\PicFillColorControl, minimum width=1.5cm, minimum height=1.2cm},
+  action/.style = {ellipse, draw, fill=\PicFillColorAction, minimum width=1cm, minimum height=1cm},
+  arrow/.style = {->, line width=1pt, draw=\PicFillColorArrow},
+  task/.style = {rectangle, draw, fill=\PicFillColorTask, text=white, inner sep=2mm},
+  pointtask/.style = {rectangle, draw, fill=\PicFillColorPT, minimum width=0.5cm, minimum height=0.5cm, text=white},
+  kernel/.style = {regular polygon, regular polygon sides=3, draw, fill=\PicFillColorKernel, text=white, font=\tiny, inner sep=1pt},
+  thread/.style = {-o, line width=.5pt}
+}
+
+\begin{tikzpicture}
+
+
+\node [main] (m1) {Main};
+\node[control, below=10mm of m1] (c1) {CP 1};
+
+\begin{pgfonlayer}{background}
+\node[control, fill=\PicFillColorLayerA, above right= -8mm and -8.6mm of c1, opacity=0.4] (cm1) {};
+\end{pgfonlayer}
+\begin{pgfonlayer}{backbackground}
+\node[control, fill=\PicFillColorLayerB, above right= -8mm and -8.6mm of cm1, opacity=0.2] (cm2) {};
+\end{pgfonlayer}
+
+\begin{pgfonlayer}{background}
+\node [main, fill=\PicFillColorLayerA, above right=-4mm and -18mm of m1] (m2) {Main};
+\draw[->, line width=1pt, draw=\PicFillColorLayerA] (m2.south) -- (cm1.north);
+\end{pgfonlayer}
+\begin{pgfonlayer}{backbackground}
[…]
