### 2025-08-12 #8756 @danielskatz
Before we get to reviewers, we need an editor, which perhaps could be Jed. But first, we need you to fix the paper...
