### 2025-08-18 9f3549c
docs: document kessel and dev-setup.sh
