+++ tutorial/standalone/flaxpy/CMakeLists.txt
@@ -7,7 +7,7 @@ cmake_minimum_required(VERSION 3.12)
 project(Flaxpy LANGUAGES CXX C)
 
 set(CXX_STANDARD_REQUIRED ON)
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
 find_package(FleCSI REQUIRED)
 
