### 2025-08-20 9cf325f
paper.md: added some middle initials
