### 2025-08-22 a5250e0
Add task launch overhead measurements
