+++ doc/sphinx/src/user-guide/performance.rst
@@ -89,3 +89,11 @@ Below presents a proper implementation of the tracing:
 
   } while(ita < max_iterations.value());
 
+In other trials using Legion in Release mode, tracing has been shown to all but completely eliminate any of the overhead from Legion's superlinear dynamic dependency analysis.
+These experiments simply used 10ms sleeps in every task and passed varying number of fields for a varying number of task launches in each trial.
+Without tracing, the asymptotic overhead time per task is approximately :math:`(560 ps)f^{2}w + (97 µs)f + (190 µs)` where f is the number of fields passed and w is legion's `window size`__.
+This represents an upper bound on the cost of Legion's dependency analysis.
+
+__ https://legion.stanford.edu/profiling/index.html
+
+By adding tracing, one is essentially reducing the window size by reducing the number of tasks Legion has to analyze. Reducing the window size manually however will not produce the same effect of reducing the task overhead dramatically and removing any nonlinearity.
