### 2025-09-03 af93f0e
ci: avoid spack-repo absolute path from other job
