### 2025-09-16 a17e814
Implemented and tested flecsi_arg().
