### 2025-09-17 11f7a91
Add noexcept for task wrappers
