### 2025-09-24 82f4a38
Repack ragged fields in ~mutator
Support copies by storing TaskBuffer in a shared_ptr
Remove param_buffers, which didn't handle vector/tuple properly
Remove other commit functions, multi_buffer, and execute_task_unbind
Remove clone, which was needed only for the param_buffers walk
Merge bind_base into bind_parameters
Destroy task parameters reliably for HPX

Remove data.cc test oudated by changes to mutator internals
Bring "Task launches" in summary.rst up to date generally
