+++ flecsi/data/test/data.cc
@@ -1,221 +1,0 @@
-#include "flecsi/data.hh"
-#include "flecsi/util/unit.hh"
-
-using namespace flecsi;
-using namespace flecsi::data;
-
-using mutator_t = typename field<std::size_t, ragged>::mutator<rw>;
-
-int
-ragged_mutator_driver(scheduler &) {
-  UNIT() {
-    // Declare backing storage and build the mutator
-    std::vector<std::size_t> memory{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11};
-    std::vector<std::size_t> endpoints{3, 9, 10, 10, 12};
-    topo::resize::Field::value_type size = 1;
-    mutator_t::TaskBuffer buffers;
-    auto m = [&memory, &endpoints, &size, &buffers]() {
-      // Declare the mutator
-      field<std::size_t, ragged>::mutator<rw> m({});
-      // Bind the backing storage for data elements
-      m.get_base().get_base().bind(util::span(memory));
-      // Bind the backing storage for the endpoints
-      m.get_base().get_offsets().bind(util::span(endpoints));
-      // Bind the backing storage for the size
-      m.get_size().get_base().get_base().bind(util::span(&size, 1));
-      // Set up overflow storage and bind to mutator
-      buffers.resize(endpoints.size());
-      m.buffer(buffers);
-      // Return the mutator
-      return m;
-    }();
-
-    // Convenience for querying because I keep having off-by-one errors
-    auto get_offset = [&endpoints](std::size_t irow) {
-      return irow == 0 ? 0 : endpoints[irow - 1];
-    };
-
-    // Test size(), capacity()
-    ASSERT_EQ(m.size(), endpoints.size());
-    for(std::size_t irow{0}; irow < m.size(); ++irow) {
-      auto row = m[irow];
-      EXPECT_EQ(row.size(), get_offset(irow + 1) - get_offset(irow));
-      EXPECT_EQ(row.capacity(), row.size());
-    }
-
-    // Test operator[]
-    std::size_t n{0};
-    for(std::size_t irow{0}; irow < m.size(); ++irow) {
-      auto row = m[irow];
-      for(std::size_t j{0}; j < row.size(); ++j) {
-        EXPECT_EQ(&row[j], &memory[n]);
-        ++n;
-      }
-    }
-
-    // Test front() and back() (no buffer)
-    for(std::size_t irow{0}; irow < m.size(); ++irow) {
-      if(irow != 3) { // m[3] is empty so front() back() would fail
-        auto const r = m[irow];
-        EXPECT_EQ(&r.front(), &r[0]);
-        EXPECT_EQ(&r.back(), &r[r.size() - 1]);
-      }
-    }
-
-    // Test empty()
-    for(std::size_t irow{0}; irow < m.size(); ++irow) {
-      EXPECT_EQ(m[irow].empty(), irow == 3); // only m[3] empty
-    }
-
-    // Test clear()
-    m[0].clear();
-    EXPECT_EQ(m[0].size(), 0);
-
-    // Test insert
-    // -- insert(iterator pos, const T & value)
-    std::size_t const v0 = 120;
-    m[0].insert(m[0].begin(), v0);
-    ASSERT_EQ(m[0].size(), 1);
[…]
