+++ flecsi/exec/CMakeLists.txt
@@ -1,6 +1,5 @@
 flecsi_headers(
   backend.hh
-  buffers.hh
   fold.hh
   fwd.hh
   kernel.hh
