### 2025-10-01 0be3837
Fix 2.4.0 release notes
