### 2025-10-01 3e790d0
Document the various task-parameter decompositions
