### 2025-10-10 8b3c720
Add quantitative costs with tracing
Replace my window-size hypothesis with discussion of its low cost
