### 2025-10-17 1cfe969
Substantially modify the HPX backend documentation
This commit implements a number of major changes to the documentation
of the HPX backend:

* Corrected spelling, punctuation, grammar, and Sphinx usage.
* Removed descriptions of HPX proper, as these can be found on the HPX
  website.
* Moved discussions of backend activities that are not specific to
  the HPX backend into the Internal Summary.
* Updated explanations of internal behavior that changed since the HPX
  backend document was initially written.
* Moved descriptions of HPX configuration parameters from the
  documentation to code com
[…]
