+++ doc/sphinx/src/developer-guide/hpx-backend.rst
@@ -1,240 +1,143 @@
-FleCSI's HPX Backend
-********************
+HPX Backend
+***********
 
-This section of the documentation is intended for developers of the core FleCSI library.
-HPX is the `C++ Standard Library for Concurrency and Parallelism <https://github.com/STEllAR-GROUP/hpx>`_. 
-For an extensive description of HPX and its API, please see the `HPX Documentation <https://hpx-docs.stellar-group.org/latest/html/index.html>`_.
+`HPX <https://hpx.stellar-group.org>`_ is an efficient user-level threading implementation that also extends the `C++ concurrency support library <https://en.cppreference.com/w/cpp/thread>`_ to operate across multiple processes.
 
-At its core, HPX is a highly efficient user-level threading implementation.
-On top of this core functionalities, HPX implements:
+Overview
+++++++++
 
-* All of the facilities related to concurrency and parallelism as defined by the C++ Standard, such as ``hpx::thread``, ``hpx::future``, or ``hpx::async``.
+The HPX backend implements data management similarly to FleCSI's MPI backend while task execution management is implemented similarly to FleCSI's Legion backend.
+Consequently, the HPX backend supports only one ``color`` per rank.
+However, it supports concurrent execution of FleCSI tasks whenever possible.
 
-* The full set of the C++17/C++20/C++23 parallel algorithms; in fact its is one of the first full openly available implementations of those.
+The HPX backend extracts execution dependencies among FleCSI tasks by analyzing the task parameters (accessor privileges) and arguments (which fields and topologies are used).
+Each of the FleCSI tasks is then scheduled as an HPX task such that it will run only after all FleCSI fields to which it has write access have been relinquished by all tasks having read and/or write access to those fields and all FleCSI fields to which it has read access have been relinquished by all tasks having write access to those fields.
+This ensures that all tasks run as early as possible and with as much concurrency as possible.
 
-  HPX adds certain extensions to the algorithms, such like asynchronous execution, parallel range based algorithms, and explicit vectorization execution policies ``simd`` and ``par_simd``.
+Dependency tracking is implemented via a set of ``hold`` objects, which track the HPX futures and active communicators that are associated with each field.
+When a task has finished execution, the corresponding future is marked as "ready".
 
-* Full set of the sender-receiver proposal (`std::execution <https://wg21.link/p2300>`_) implemented using C++17.
+As communicator creation is expensive, the HPX backend strives to reuse existing communicators.
+Whether or not a task requires a communicator—as one of its dynamic descendants might—it begins by walking the dependency DAG to prune nodes (``comms`` objects—one per task) that no longer are needed, thereby simplifying the graph for subsequent traversals.
+The task then adds a single node to the DAG, unless it has nowhere to store it.
+This node points to all direct predecessors, "pointing through" any that lack communicators to the nearest indirect predecessor that does not.
+Finally, the task stores a shared pointer to the node in appropriate fields.
 
-* An uniform integration of your Kokkos, CUDA, HIP, and SYCL (oneAPI) kernels enabling their asynchronous invocation.
+As a result of these graph operations, communicators effectively are moved down the DAG to the deepest point where they still can be discovered by potential users of those communicators.
+A communicator held by a task with no dependents cannot be migrated.
+Rather, the communicator remains inaccessible until it is destroyed along with the regions used by the task graph when the computation performed by that graph completes.
 
-* Fully distributed operation enabling remote management of distributed C++ objects and the invocation of global C++ functions and C++ member functions on those objects. 
[…]
