### 2025-10-20 acd9c94
Add task specialization documentation
