+++ doc/sphinx/src/user-guide/execution.rst
@@ -1,21 +1,20 @@
 Execution Model
 ***************
 
-This section describes the FleCSI execution model, which is based on the concepts of tasks and kernels.
-These abstractions allow users to express parallelism at different levels, from task scheduling to fine-grained data access.
+This section describes the FleCSI execution model.
+FleCSI expresses parallelism via tasks, which are coarse-grained and can be distributed, and kernels, which are fine-grained and utilize shared memory.
 
 Schedulers
 ++++++++++
 
 A FleCSI *scheduler* is an object that manages the launch of tasks.
-Where and when to execute the task is determined based on the task's parameter.
-They select the appropriate execution space based on the task and platform.
+Where and when to execute the task is determined based on the task's parameter types and arguments.
 
-Schedulers are obtained from a *control_point* when writing an action inside the control model:
+Schedulers are obtained from the control-model object provided to an action:
 
 .. code-block:: c++
 
-   void action(control_point &cp) {
+   void action(control_policy &cp) {
      flecsi::scheduler s = cp.scheduler();
      s.execute<task>(...);
    }
@@ -26,14 +25,13 @@ Tasks
 A task in FleCSI serves as the bridge between two core parts of the execution model.
 On the caller side, each ``process``, also referred to as a ``rank`` in the MPI terminology, executes its own copy of the ``main`` function and the associated control model.
 
-Each process executes its own instance of ``main`` and the control model, and so all processes collectively launch and schedule a task based on its :ref:`field privileges <field-accessors>`.
-The function for a task then executes, typically several times concurrently.
+After all processes collectively launch a task, its function then executes, typically several times concurrently.
 Each of these executions is referred to as a *point task*; their number does not necessarily match the number of processes, and they can run on any process.
 This allows FleCSI to support flexible and scalable task-parallel execution.
 
 .. figure:: tikz/flecsi_execution_model.svg
 
-   The different levels of abstraction in FleCSI.
+   The different levels of execution in FleCSI.
 
 Callee Side
 ^^^^^^^^^^^
@@ -47,15 +45,14 @@ For instance, the task ``foo`` may be defined as:
      flecsi::field<double>::accessor<ro, ro> f1,
      flecsi::field<std::size_t>::accessor<rw, ro> f2) noexcept { /* ... */ }
 
-Tasks must be declared ``noexcept`` because they are scheduled and executed indirectly and cannot throw exceptions.
-The first parameter, ``e``, specifies the execution space where the task runs.
-Available options include `cpu`, `omp`, `gpu`, or `accelerator`.
-The `accelerator` space is flexible and will use the backend selected when compiling Kokkos (e.g., `Serial`, `OpenMP`, `CUDA`, etc.).
-This allows users to target different hardware execution spaces without modifying the task code.
-If the execution space is not explicitly specified in the callee declaration, the task will default to CPU execution.
+Tasks must be declared ``noexcept``: they are called asynchronously without any means of handling an exception.
+Here the first parameter, ``e``, specifies the execution space where the task runs.
+Available options include ``cpu``, ``omp``, ``gpu``, or ``accelerator``.
+The ``accelerator`` space is flexible and will use the backend selected when compiling Kokkos (e.g., CUDA, OpenMP, or serial), allowing the same task code to use different kinds of hardware.
+If there is no execution space parameter, the task will default to CPU execution.
 
-The field references provided during the task launch are automatically converted into :ref:`field-accessors`.
-These accessor types grant access to the real data, potentially after hidden memory copies or device transfers, depending on the privileges set by the user.
[…]
