+++ api/core/structflecsi_1_1exec_1_1task__prologue.js
@@ -1,6 +1,0 @@
-var structflecsi_1_1exec_1_1task__prologue =
-[
-    [ "task_prologue", "structflecsi_1_1exec_1_1task__prologue.html#a55f6f418081ec894921fe98a3f7df82d", null ],
-    [ "visit", "structflecsi_1_1exec_1_1task__prologue.html#acdf1fd0f60010b2525b106c5223a1d50", null ],
-    [ "visit", "structflecsi_1_1exec_1_1task__prologue.html#a8f8bdb329553f7edea9da2429f54eb6c", null ]
-];
