### 2025-10-22 72434f7
Use normal tasks in narray
Call meta_data::ghosts repeatedly to avoid communication
Remove idx_itvls; move set_dests and set_ptrs from base to use meta_data
Reduce index_definition::process_colors to one-color decode
Remove thus unused topo::concatenate
