### 2025-10-23 8af440f
Add scheduler::wait
