### 2025-10-23 984dcf3
Don't use "rank" as synonym for "process"
Make minor cleanups in immediate vicinity of changes
