### 2025-10-23 4456ef2
Remove unused |br| declarations
