+++ doc/sphinx/src/tutorial/poisson.rst
@@ -1,7 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
 Poisson
 *******
 
