+++ doc/sphinx/src/tutorial/flaxpy.rst
@@ -1,7 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
 DAXPY
 *****
 
