+++ doc/sphinx/src/tutorial.rst
@@ -1,7 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
 Tutorial
 ********
 
