+++ doc/sphinx/src/tutorial/parallel.rst
@@ -1,8 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
-
 Shared-memory parallelism
 *************************
 
