+++ doc/sphinx/src/tutorial/control.rst
@@ -1,7 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
 Control Model
 *************
 
