+++ doc/sphinx/src/tutorial/ntree.rst
@@ -1,7 +1,3 @@
-.. |br| raw:: html
-
-   <br />
-
 N-Tree
 ******
 
