### 2025-10-23 6058ed3
Use `term` for definitions
