+++ doc/sphinx/src/developer-guide/summary.rst
@@ -11,16 +11,16 @@ Components
 
 FleCSI performs only a few fundamental actions, each of which corresponds to one of the namespaces/directories under ``flecsi``:
 
-* ``data``: Allocate distributed one-dimensional arrays of arbitrary trivial types (*fields*).
-* ``topo``: Organize those arrays into *topologies* that represent computational physics domains of several kinds.
-* ``exec``: Call user-specified functions (*tasks*) with pointers to the allocated memory (in the form of *accessors*).
+* ``data``: Allocate distributed one-dimensional arrays of arbitrary trivial types (`fields`).
+* ``topo``: Organize those arrays into `topologies` that represent computational physics domains of several kinds.
+* ``exec``: Call user-specified functions (`tasks`) with pointers to the allocated memory (in the form of `accessors`).
 * ``io``: Save and restore the contents of fields to and from disk.
 * ``flog``: Aggregate diagnostic output from multiple processes.
 
 Some FleCSI actions have a single implementation while others rely on different implementations depending on the underlying communication and tasking mechanisms.
 In the preceding list, the ``topo`` and ``flog`` components have a single implementation.
-The other FleCSI components are built atop *backends* that leverage some external mechanism for allocating memory, transferring distributed data, and executing tasks.
-The common backend API comprises a small set of classes and function templates that are sufficient to implement the rest of FleCSI; each class/function template is called an *entry point*.
+The other FleCSI components are built atop `backends` that leverage some external mechanism for allocating memory, transferring distributed data, and executing tasks.
+The common backend API comprises a small set of classes and function templates that are sufficient to implement the rest of FleCSI; each class/function template is called an `entry point`.
 Most entry points are defined in files named ``policy.hh`` in a backend-specific directory in each of the ``data``, ``exec``, and ``io`` components.
 
 A backend is selected at CMake configuration time by providing a value for the ``FLECSI_BACKEND`` option.
@@ -58,7 +58,7 @@ Generally, ``exec`` is higher than ``data`` because launching a task involves ob
 Finally, ``util`` is the lowest, generic level.
 
 There is also an expected hierarchy beyond FleCSI.
-In particular, it is expected that the topology *specializations* required as policies for the core topology *categories* (*e.g.*, to specify the number of spatial dimensions) are provided by intermediate clients like FleCSI-sp that also provide domain-specific interfaces to application code.
+In particular, it is expected that the topology `specializations` required as policies for the core topology `categories` (*e.g.*, to specify the number of spatial dimensions) are provided by intermediate clients like FleCSI-sp that also provide domain-specific interfaces to application code.
 (Topology specializations are not C++ explicit or partial specializations, although they may be generated by a class template.)
 
 Compound clients
@@ -79,23 +79,23 @@ Task separation
 ^^^^^^^^^^^^^^^
 
 Field data appears as actual, borrowed arrays in a task but as owning, opaque handles in the caller.
-As such, certain *caller-only* or *task-only* types are used only in one context or the other, and the types of some task arguments differ from the types of the parameters they indirectly initialize.
+As such, certain `caller-only` or `task-only` types are used only in one context or the other, and the types of some task arguments differ from the types of the parameters they indirectly initialize.
 Conveniently, this separation also guarantees thread safety in the event that tasks execute concurrently with their callers in the same process.
 
 Storage
 ^^^^^^^
 
 The FleCSI data model is based on `that of Legion <https://legion.stanford.edu/tutorial/logical_regions.html>`_, with certain simplifications.
-Each field has a value at each of several *index points*, each of which is a pair of abstract integers.
-The first indicates the *color* to which the index point pertains, which determines the execution agent that typically operates on the associated field values.
[…]
