### 2025-10-24 872b311
Remove statements accidentally about the future
