### 2025-10-28 bd5f52b
Label Runtime Model in figure
