### 2025-10-28 e109c00
Split Control Model User Guide page from tutorial
Move the relevant images
