+++ doc/sphinx/user/control.rst
@@ -1,0 +1,179 @@
+Control Model
+*************
+
+The FleCSI control model allows users to define the high-level structure
+of an application
+using a control-flow graph (CFG) of `control points`,
+under each of which a directed acyclic graph (DAG) of `actions` can be defined.
+Actions in turn launch `tasks` to operate on distributed data.
+
+.. sidebar:: Control-model concepts
+
+   **Control Points** are the logical steps in a control-flow graph.
+   These are not actual execution units. However, they define named
+   anchors where actions can be registered that will be executed when
+   the control-flow graph reaches that control point.
+
+   **Actions** are C++ functions that are executed under a control point.
+
+   **Tasks** are C++ functions launched by an action
+   and that access distributed data.
+
+For FleCSI developers, the control model replaces the normal hard-coded
+execution structure of an application, instead providing a well-defined,
+extensible mechanism which can easily be verified and visualized.
+
+Consider the following example of a traditional *implicit* control
+model:
+
+.. code-block:: cpp
+
+   int main(int argc, char ** argv) {
+     int step{0}, steps{0};
+
+     // initialization stage
+     initialize(steps);
+
+     while(step<steps) {
+       // simulation stages
+       advanceA();
+       advanceB();
+       advanceC();
+
+       // analysis stages
+       analyzeA();
+       analyzeB();
+       analyzeC();
+     } // while
+
+     // finalization stage
+     finalize();
+   } // main
+
+While this example is straightforward, it is also restrictive because it
+defines a single point in the application where all stages must be
+known.
+Suppose that an advanced user would like to add an *advance* or
+*analysis* stage to the simulation, e.g, to test a new algorithm or to
+customize their metrics?
+With the traditional model above, she would need to have access to the
+code base and would have to rewrite the main simulation loop with the
+new execution logic.
+This is error prone, and worse it necessarily forces the code to
+diverge (unless every alteration to the application is vetted and
+merged).
+
+FleCSI's control model offers a much cleaner solution, which is both
+easier for the user and safer for the core application developers.
+
+----
+
+Control-Flow Graphs & Directed Acyclic Graphs
++++++++++++++++++++++++++++++++++++++++++++++
+
+:numref:`cfg` shows a complete representation of a FleCSI control model.
+This figure was actually generated by a FleCSI example application using
+the ``control::write_graph`` function, which outputs a dot file of the control
+model.
+The white control-point nodes in the figure (labeled: *Control Point X*)
+define a control-flow graph. Notice that the control points do not form
+a DAG, i.e., they cycle (indicated by the dashed line from *Control
+Point 3* to *Control Point 2*).
+This is true in general and allows users to create complex looping
+structures within their applications.
+
+At each control point in the CFG, there are one or more actions defined:
+e.g., under *Control Point 2* are the actions *A* through *E*.
+The actions are where the actual work of the simulation is executed.
+
+.. important::
+
+   The actions in a FleCSI control model are not tasks! The control
+   model capability in FleCSI is orthogonal to Legion's tasking model;
+   i.e., they are complementary. **FleCSI actions are functions that
+   execute tasks!** When the Legion backend to FleCSI is in use, the
+   control model exposes a sequential program order to the Legion
+   runtime that can be analyzed for data dependencies to increase
+   concurrency and allow load balancing during execution.
+
+.. MAINTAINER NOTE: The image for this figure was generated by the
+   control-overview program that is located in
+   'flecsi/tutorial/.auxiliary/control-overview'.
+.. _cfg:
+.. figure:: images/overview.svg
+   :align: center
+   :width: 90%
+
+   Example FleCSI Control Model.
+
+.. sidebar:: Sequential Ordering
+
+   .. _sequential:
+   .. figure:: images/sequential.svg
+      :align: center
+      :width: 50%
+
[…]
