### 2025-10-28 4f2caa6
Reorder User Guide
Put cross-cutting topics last
