### 2025-10-28 6680ea2
Simplify Sphinx file names
Inline src subdirectory
Remove "-guide" from directory names
