+++ doc/sphinx/tutorial/runtime.rst
@@ -25,7 +25,7 @@ This example demonstrates a minimal use of FleCSI that just executes an
 action to print out *Hello World*. Code for this example can be found in
 *tutorial/1-runtime/1-minimal.cc*.
 
-.. literalinclude:: ../../../../tutorial/1-runtime/1-minimal.cc
+.. literalinclude:: ../../../tutorial/1-runtime/1-minimal.cc
   :language: cpp
 
 ----
@@ -99,7 +99,7 @@ Declaring Options
 Let's consider the first *Car Options* option: ``--level``. To declare
 this option, we use the following declaration:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add an integer-valued command-line option with a default value
   :end-before: // Add a string-valued command-line option with a default value
@@ -143,7 +143,7 @@ parameters:
 The next option ``--transmission`` is similar but uses a ``std::string``
 value type:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add a string-valued command-line option with a default value
   :end-before: // Add an option that defines an implicit value.
@@ -154,7 +154,7 @@ std::string) the default value is also a string.
 The last option in the "Car Options" section ``--child-seat``
 demonstrates the use of flecsi::option_implicit:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add an option that defines an implicit value.
   :end-before: // Add a an option to a different section,
@@ -177,7 +177,7 @@ The first option in the *Ride Options* section ``--purpose`` takes an
 integer value *0* or *1*. This option is declared with the following
 code:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add a an option to a different section,
   :end-before: // Add an option with no default.
@@ -192,7 +192,7 @@ an implicit value and zero values (meaning that it takes no values). The
 the flag is passed.  This will be useful to demonstrate how we can check
 whether or not an option was passed in the next section:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add an option with no default.
   :end-before: // Add a positional option. 
@@ -200,7 +200,7 @@ whether or not an option was passed in the next section:
 The final option in this example is a positional option: i.e., it is an
 argument to the program itself.
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
   :start-at: // Add a positional option. 
   :end-before: // User-defined program options are available after
@@ -220,7 +220,7 @@ line and on its default and implicit values.
 
 Options that have a default value defined do not need to be tested:
 
-.. literalinclude:: ../../../../tutorial/1-runtime/2-program_options.cc
+.. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
   :language: cpp
[…]
