+++ doc/sphinx/tutorial/data.rst
@@ -18,7 +18,7 @@ Global fields are used to store global variables/objects that can be
 accessed by any task.  For the common case of only one value for each field, it
 is natural to use the ``data::single`` layout.
 
-.. literalinclude:: ../../../../tutorial/4-data/1-global.cc
+.. literalinclude:: ../../../tutorial/4-data/1-global.cc
   :language: cpp
   :start-at: template<typename T>
   :end-at: const single<double>::definition<global> gfield;
@@ -28,7 +28,7 @@ In general, a coloring describes the structure of a topology and its distributio
 The global topology is a special case that does not actually use colors; its "coloring" is simply a count of values for each field.
 Writing to a global field requires a single task launch.
 
-.. literalinclude:: ../../../../tutorial/4-data/1-global.cc
+.. literalinclude:: ../../../tutorial/4-data/1-global.cc
   :language: cpp
   :start-at: void
   :end-at: // advance()
@@ -42,7 +42,7 @@ Example 2: Index data
 
 A field on an ``index`` topology stores one value for each color.
 
-.. literalinclude:: ../../../../tutorial/4-data/2-index.cc
+.. literalinclude:: ../../../tutorial/4-data/2-index.cc
   :language: cpp
   :start-at: using namespace flecsi;
   :end-at: // advance()
@@ -53,7 +53,7 @@ Example 3: Dense data
 A dense field is a field defined on a dense topology index space.  In
 this example we allocate a ``pressure`` field on the ``cells`` index space of the ``canonical`` topology.
 
-.. literalinclude:: ../../../../tutorial/4-data/3-dense.cc
+.. literalinclude:: ../../../tutorial/4-data/3-dense.cc
   :language: cpp
   :start-at: const field<double>::definition<canon, canon::cells> pressure;
   :end-at: const field<double>::definition<canon, canon::cells> pressure;
@@ -66,7 +66,7 @@ Here, a file is the source of the mesh (for purposes of illustration).
 The resulting coloring is used to initialize two meshes ``canonical`` and ``cp``, and the ``copy`` task operates on both of them at once using a low-level accessor.
 The ``init`` and ``print`` tasks, by contrast, use a `topology accessor` as a parameter that provides access to the structure of the mesh via the ``entities`` function.
 
-.. literalinclude:: ../../../../tutorial/4-data/3-dense.cc
+.. literalinclude:: ../../../tutorial/4-data/3-dense.cc
   :language: cpp
   :start-at: void
   :end-at: // advance()
@@ -76,7 +76,7 @@ Example 4: Ragged data
 A ragged field stores a variable amount of data at each index point.
 It is defined in much the same way as a dense field, but using it involves additional steps.
 
-.. literalinclude:: ../../../../tutorial/4-data/4-ragged.cc
+.. literalinclude:: ../../../tutorial/4-data/4-ragged.cc
   :language: cpp
   :start-at: ints
   :end-at: rag;
@@ -85,14 +85,14 @@ Because the total storage required for a ragged field is not determined by the s
 The field reference for a ragged field has a special interface for managing it.
 Here we use a simple helper task (called ``allocate``) with that interface to specify a fixed, known total, although typically a heuristic overallocation is required:
 
-.. literalinclude:: ../../../../tutorial/4-data/4-ragged.cc
+.. literalinclude:: ../../../tutorial/4-data/4-ragged.cc
   :language: cpp
   :start-at: predefined
   :end-at: }
 
 After executing a task to store the sizes, they must be applied with ``resize``:
 
-.. literalinclude:: ../../../../tutorial/4-data/4-ragged.cc
[…]
