+++ doc/sphinx/user.rst
@@ -1,0 +1,18 @@
+User Guide
+**********
+
+.. toctree::
+  :caption: User Guide:
+
+  user/programming
+  user/data
+  user/topologies
+  user/execution
+  user/kernels
+  user/specializations
+  user/utilities
+  user/performance
+  user/profiling
+  user/cmake
+
+.. vim: set tabstop=2 shiftwidth=2 expandtab fo=cqt tw=72 :

[…]
