+++ doc/sphinx/user.rst
@@ -1,8 +1,9 @@
 User Guide
 **********
+These pages constitute a reference for the concepts used in FleCSI.
+They do not mention every component of the interface as does the :doc:`api`, nor contain extensive examples as does the :doc:`tutorial`, but instead provide context beyond the scope of either.
 
 .. toctree::
-  :caption: User Guide:
 
   user/control
   user/data
@@ -15,5 +16,3 @@ User Guide
   user/programming
   user/performance
   user/profiling
-
-.. vim: set tabstop=2 shiftwidth=2 expandtab fo=cqt tw=72 :
