### 2025-10-28 6f28520
Add Runtime User Guide page
Reuse layers-of-execution figure; move it to top of Execution Model too
