### 2025-10-28 a70fb0e
Discuss single and index launches in User Guide
