### 2025-10-29 946b46c
Explain the use of data::multi objects
