### 2025-10-29 8927c9c
Expand future discussion
Mention use as parameters not just in Performance
