### 2025-10-29 8a62829
Discuss MPI-task argument/return value semantics
