### 2025-10-29 2606114
Remove redundant material from Internal Summary
