+++ doc/sphinx/developer/summary.rst
@@ -61,17 +61,6 @@ There is also an expected hierarchy beyond FleCSI.
 In particular, it is expected that the topology `specializations` required as policies for the core topology `categories` (*e.g.*, to specify the number of spatial dimensions) are provided by intermediate clients like FleCSI-sp that also provide domain-specific interfaces to application code.
 (Topology specializations are not C++ explicit or partial specializations, although they may be generated by a class template.)
 
-Compound clients
-^^^^^^^^^^^^^^^^
-
-FleCSI's design is meant to support a client simulation program that comprises a number of subsystems with minimal coordination among them.
-For example, fields are registered on merely a specialization, endowing all instances of its topology type with the field in question without requiring changes to its definition.
-(Memory is allocated only if the field is used, so clients can err on the side of registering whatever fields might be useful.)
-Additionally, the registrations take the form of non-local variables such that any source file in the program can contribute them independently.
-
-Another example is the callback-based control flow system, which allows actions (and ordering constraints) to be associated with each part of the overall simulation cycle structure.
-Again, each such action is a non-local variable that can be supplied separately anywhere in the program.
-
 Data
 ++++
 
@@ -169,22 +158,10 @@ The class template ``coloring_slot``, defined in ``coloring.hh`` automates invok
 Access
 ^^^^^^
 
-Access to a field is requested with a caller-only `field reference` which identifies a field and a topology instance.
-A field reference may be passed as an argument for an accessor parameter of a task.
-Accessors are task-only; their types are usually spelled ``field<T,L>::accessor<P,...>``, where each ``P`` is a `privilege` that specifies read and write permissions for some part of the field.
-
-When a single privilege is supplied (``field<T,L>::accessor<P>``), it refers to `owned` index
-points.  For two privileges (``field<T,L>::accessor<P1,P2>``), the first (``P1``) refers to owned
-index points and the second (``P2``) to `ghosts` that may be copied automatically from
-pre-specified shared points.  For three privileges (``field<T,L>::accessor<P1,P2,P3>``), the first
-(``P1``) refers to `exclusive` index points, the second (``P2``) to `shared` index points, and the
-third (``P3``) to ghosts.
-Ghost copies are performed only when ghosts are read and not written and shared points have been written more recently than the previous read or write.
-(There is no mechanism at present to overlap the ghost copies with a task that does not require access to ghosts or write access to shared points.)
-
 Internally, all of the (variable number of) privileges for an accessor are combined into a `privilege pack`.
 The syntax ``field<T,L>::accessor1<P>`` may be used to reuse such a pack.
 The actual type of an accessor is ``data::accessor<L,T,P>``, which must be used for template argument deduction to apply.
+Note that there is no mechanism at present to overlap the ghost copies with a task that does not require access to ghosts or write access to shared points.
 
 When the arrays for one or more index points in a ``ragged`` or ``sparse`` field are resized, they must be repacked.
 To do so efficiently, the interface for such resizing is provided by accessor variants called `mutators`, which use temporary storage (from the ordinary heap) to track changes made by a task and then apply those changes when it finishes.
@@ -202,11 +179,6 @@ Because the structural information about a topology is often necessary for using
 Topology accessors are of course also task-only; a topology accessor parameter is matched by a ``topology`` argument.
 The topology's ``access`` type is used wrapped in the ``topology_accessor`` class template defined in ``topology_accessor.hh``.
 
-`Multi-color accessors` allow an execution agent to access data outside of its color (beyond that supplied by ghost copies), including the special case of data outside of the data model altogether (*e.g.*, distributed objects created by MPI-based libraries).
-These take the form of a sequence of color-accessor pairs inside a task; the accessor can also be a mutator or a topology accessor.
[…]
