### 2025-10-29 2dc61f6
Fix flushing motivation
