### 2025-10-29 5a95d17
Move accessor description to first use
