### 2025-10-29 7dcd16f
Rename kernels.rst to match tutorial
Deduplicate them against each other
