### 2025-10-29 417f1dd
Reduce launch_domain discussion
