### 2025-10-29 f6ff613
Fix overreach in project summary
