+++ doc/sphinx/user/control.rst
@@ -59,7 +59,7 @@ customize their metrics?
 With the traditional model above, she would need to have access to the
 code base and would have to rewrite the main simulation loop with the
 new execution logic.
-This is error prone, and worse it necessarily forces the code to
+This is error-prone, and worse it necessarily forces the code to
 diverge (unless every alteration to the application is vetted and
 merged).
 
@@ -73,9 +73,9 @@ Control-Flow Graphs & Directed Acyclic Graphs
 
 :numref:`cfg` shows a complete representation of a FleCSI control model.
 This figure was actually generated by a FleCSI example application using
-the ``control::write_graph`` function, which outputs a dot file of the control
+the ``control::write_graph`` function, which outputs a Dot file of the control
 model.
-The white control-point nodes in the figure (labeled: *Control Point X*)
+The white control-point nodes in the figure (labeled *Control Point X*)
 define a control-flow graph. Notice that the control points do not form
 a DAG, i.e., they cycle (indicated by the dashed line from *Control
 Point 3* to *Control Point 2*).
@@ -96,9 +96,7 @@ The actions are where the actual work of the simulation is executed.
    runtime that can be analyzed for data dependencies to increase
    concurrency and allow load balancing during execution.
 
-.. MAINTAINER NOTE: The image for this figure was generated by the
-   control-overview program that is located in
-   'flecsi/tutorial/.auxiliary/control-overview'.
+.. The control model is that of the "cycle" test.
 .. _cfg:
 .. figure:: images/overview.svg
    :align: center
