+++ doc/sphinx/tutorial/runtime.rst
@@ -36,18 +36,17 @@ Example 2: Program Options
 FleCSI supports a program options capability based on `Boost Program
 Options`__ to simplify the creation and management of user-defined
 command-line options. The basic syntax for adding and accessing program
-options is semantically similar to the Boost interface (You can find
-documentation using the above link.) However, there are some notable
+options is semantically similar to the Boost interface, but there are some notable
 differences:
 
-* FleCSI internally manages the boost::program_options::value variables
-  for you, using boost::optional.
+* FleCSI internally manages the ``boost::program_options::value`` variables
+  for you, using ``boost::optional``.
 
 * Positional options are the mechanism that should be used for
   *required* options.
 
 * Default, implicit, zero, and multi value attributes are specified in
-  the flecsi::program_option constructor as an std::initializer_list.
+  the ``flecsi::program_option`` constructor as a ``std::initializer_list``.
 
 This section of the tutorial provides examples of how to use FleCSI's
 program option capability.
@@ -57,9 +56,8 @@ __ https://www.boost.org/doc/libs/release/doc/html/program_options.html
 Example Program
 ^^^^^^^^^^^^^^^
 
-In this example, imagine that you have a program that takes information
-about a taxi service (The options are silly and synthetic. However they
-demonstrate the basic usage of the flecsi::program_option type.) The
+In this example, imagine that you have a program that takes information about a taxi service (the options are silly and synthetic, but they demonstrate the basic usage of ``flecsi::program_option``).
+The
 command-line options and arguments for the program allow specification
 of the following: trim level, transmission, child seat, purpose
 (personal or business), light speed, and a passenger list. The first two
@@ -96,7 +94,8 @@ Declaring Options
   FleCSI program options must be created before calling ``flecsi::initialize`` (and must survive through all uses of their value).
   It is often convenient to declare them in a namespace in a header file (in which case, they must also be declared ``inline``).
 
-Let's consider the first *Car Options* option: ``--level``. To declare
+Let's consider the first *Car Options* option, ``--level``.
+To declare
 this option, we use the following declaration:
 
 .. literalinclude:: ../../../tutorial/1-runtime/2-program_options.cc
@@ -105,37 +104,36 @@ this option, we use the following declaration:
   :end-before: // Add a string-valued command-line option with a default value
 
 First, notice that the flecsi::program_option type is templated on the
-underlying option type *int*. In general, this can be any valid C++
-type.
+underlying option type ``int``.
 
-This constructor to flecsi::program_option takes the following
+This constructor to ``flecsi::program_option`` takes the following
 parameters:
 
-* *section ("Car Options")*: |br|
+* ``section`` (``"Car Options"``): |br|
   Identifies the section. Sections are generated automatically, simply
   by referencing them in a program option.
 
-* *flag ("level,l")*: |br|
+* ``flag`` (``"level,l"``): |br|
   The long and short forms of the option. If the string contains a
   comma, it is split into *long name,short name*. If there is no comma,
   the string is used as the long name with no short name.
 
-* *help ("Specify...")* |br|
+* ``help`` (``"Specify..."``) |br|
   The help description that will be displayed when the usage message
   is printed.
 
-* *values ({{flecsi::option_default, ...}})* |br|
+* ``values`` (``{{flecsi::option_default, ...}}``) |br|
   This is a
-  std::initializer_list<flecsi::program_option::initializer_value<int>>.
-  The possible values are flecsi::option_default,
[…]
