### 2025-10-30 8c573b9
Document topo::id
It is mentioned in many places already
