### 2025-10-30 b45c8f7
Describe ghost privileges briefly in tutorial
