### 2025-10-30 aa6d9a7
Remove redundant context functions
Make the relevant data members private
