+++ flecsi/run/mpi/context.cc
@@ -12,7 +12,7 @@ context_t::start(const std::function<int()> & action, bool) {
   context::start();
 
   context::threads_per_process_ = 1;
-  context::threads_ = context::processes_;
+  threads_ = processes();
 
   return detail::data_guard(), task_local_base::guard(), action();
 }
