### 2025-11-01 #9333 @jedbrown
Hi @lindsayad @mprobson 👋 Welcome to JOSS and thanks for agreeing to review! The comments from @editorialbot above outline the review process, which takes place in this thread (possibly with issues filed in the [FleCSI repository](https://github.com/flecsi/flecsi)). I'll be watching this thread if you have any questions.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention this issue so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within a month or two. Please let me know if you require some more time. We can also use [editorialbot](https://joss.readthedocs.io/en/latest/editorial_bot.html#reminding-reviewers-and-authors) to set automatic reminders if you know you'll be away for a known period of time.
