### 2025-11-05 80939c4
cmake: switch to C++20
