### 2025-11-06 flecsi/flecsi#5 @lindsayad
Configure/build documentation doesn't mention possible need for device compiler specification
This is my configuration summary
```
#------------------------------------------------------------------------------#
# Configuration Summary
#------------------------------------------------------------------------------#

  FleCSI Version: 2 (feature branch)

  Repository state: 11 commits since v2.4.1 (f57a634e3f1f136e8932ad81f267d3b69657ae15)

  CMAKE_INSTALL_PREFIX: /data/lindad/flecsi/installed

  CMAKE_CXX_COMPILER: /usr/bin/c++
  CMAKE_CXX_COMPILER_VERSION: 13.3.0
  CMAKE_C_COMPILER: /usr/bin/cc
  CMAKE_C_COMPILER_VERSION: 13.3.0

  FLECSI_BACKEND: mpi
  ENABLE_FLOG: ON
  ENABLE_UNIT_TESTS: OFF

  ENABLE_GRAPHVIZ: OFF
  ENABLE_HDF5: OFF
  CALIPER_DETAIL: none

  Kokkos_DEVICES: OPENMP;SERIAL;CUDA
```
When building I encounter
```
c++: error: unrecognized command-line option ‘-Wext-lambda-captures-this’
c++: error: unrecognized command-line option ‘-arch=sm_89’
```
`nvcc` would know what to do with these but the host compiler doesn't

Ref https://github.com/openjournals/joss-reviews/issues/9333
