### 2025-11-06 flecsi/flecsi#5 @lindsayad
I'm assuming this crops up because I configured Kokkos with cuda but I don't know for sure
