### 2025-11-08 flecsi/flecsi#5 @rbberger
> C++17 compliant compiler

@lindsayad NVCC is not such a compiler. Kokkos uses that by default when configuring it for CUDA. However, since we need a real C++ compiler, you need to configure Kokkos (and therefore FleCSI) to use LLVM/Clang as C++ and CUDA compiler.

(we should be more explicit about the compiler constraints though)
