### 2025-11-12 92f8f6c
Generalize "point task" where appropriate
