### 2025-11-12 231e38e
Clarify non-meaning of particle index points
