### 2025-11-14 55eafa0
Fix MPI task conversions, again
Apply conversions regardless of backend
Remove template parameter for disabling them
Trap non-movable parameters
Explain their implementation issues in Summary
Fix orphaned sentence there
