### 2025-11-18 d3decd1
Add synchronous tasks
Add task_class member and internal task_attributes_mask_t value
Generalize more error messages and prose about MPI tasks
