+++ doc/sphinx/user/programming.rst
@@ -20,7 +20,9 @@ Whether or not the MPI backend is in use, a FleCSI application is an MPI program
 The control model actions run serially on each process and must perform the same sequence of collective calls into FleCSI with the same arguments.
 (In certain cases, it is the identity rather than the value of the arguments that matters; for example, a mesh coloring might be distributed (rather than replicated) over multiple processes, but that distributed object is the same object for the purpose of initializing a topology.)
 However, it is unspecified on which processes task instances execute, so it is not generally meaningful to modify data outside the task.
-Moreover, tasks are asynchronous: ``scheduler::execute`` may return before they complete and task instances from multiple task launches may run out of order or in parallel.
+Moreover, tasks are normally asynchronous: ``scheduler::execute`` may return before they complete and task instances from multiple task launches may run out of order or in parallel.
+To avoid dangling references, arguments passed to a task are immediately used to initialize objects of the task's decayed parameter types.
+(``std::thread`` does something similar with the decayed *argument* types.)
 
 The threads necessary to implement this impose the ordinary responsibility of thread safety among tasks as well as between them and the actions.
 Because the threads may be pooled, they provide only the `parallel forward progress guarantee <https://en.cppreference.com/w/cpp/language/memory_model#Parallel_forward_progress>`_ (invalidating certain collective operations).
@@ -33,6 +35,12 @@ Optional Semantics
 ++++++++++++++++++
 Stronger semantics for tasks are available at a performance cost.
 
+Synchronous Tasks
+^^^^^^^^^^^^^^^^^
+The caller of any task can immediately call ``wait`` on its future (with known performance costs), but if it is known statically that the caller will wait it is possible to initialize task parameters directly from task arguments (even though they might be temporary objects).
+Certain copies can thus be avoided without the syntactic overhead of using pointers.
+Furthermore, a synchronous task can accept references to non-movable types, but only if they bind directly to the arguments (without a user-defined conversion, in the current implementation).
+
 MPI Groups
 ^^^^^^^^^^
 If a task accepts a ``group::match`` as (part of) a parameter, its point tasks are assigned to processes (according to MPI rank).
@@ -42,6 +50,9 @@ This control is necessary for interprocess communication but also has ancillary 
 #. Tasks can accept pointers to non-const objects (without aliasing) and non-copyable objects by value.
 #. The various calling processes can provide different argument values for non-FleCSI types.
 #. Since no data relocation can be needed between two such tasks, fields used *only* by such tasks in one memory space may use non-trivial data types.
+
+If the task is also synchronous, such that argument lifetimes are not a concern, its parameters can further be any kind of reference.
+If such a task immediately follows a call to ``scheduler::wait``, no other point tasks can run concurrently, so global data can be modified without race conditions.
 
 Concurrent Tasks
 ^^^^^^^^^^^^^^^^
@@ -54,9 +65,5 @@ This ordering can also reduce parallelism if the tasks use different fields; tas
 
 MPI Tasks
 ^^^^^^^^^
-The execution of ``mpi`` tasks (regardless of backend) provides even stronger semantics:
-
-#. Because they run synchronously, their parameters may be references to non-const objects.
-#. Because additionally no other point tasks are executed concurrently with them, they can access global data without race conditions.
-
-However, their return values are processed in the normal fashion and must be trivially relocatable.
+An ``mpi`` task (regardless of backend) has all the properties above (including an implicit call to ``scheduler::wait``).
+Note that return values are still processed in the normal fashion and must be trivially relocatable.
