### 2025-11-19 9b65e25
Fix Poisson io task
