### 2025-11-19 8e5f862
format: apply clang-format 20.1.8
