### 2025-11-19 4e8cea6
Generalize layout.hh to tags.hh
Collect tag types from launch.hh and topology_slot.hh
