### 2025-11-25 a7e2509
Factor out exec::launch
Use it in task_wrapper, which reduce to a function
Generalize HPX's cv-qualifier removal; lose its bad annotation scope
Remove impossible static assertions
