### 2025-12-02 e3be189
Convert vector elements
Preserve element type when possible
