### 2025-12-05 6bfe43a
Move flushing logic into flog
