### 2025-12-09 10e44ca
Remove FLOG_ENABLE_MPI option
Flog meets its specification only with it set
