+++ flecsi/config.hh.in
@@ -68,7 +68,6 @@ Format: \code(major << 16) | (minor << 8) | patch\endcode
 #define FLOG_STRIP_LEVEL @FLOG_STRIP_LEVEL@
 #cmakedefine01 FLOG_ENABLE_COLOR_OUTPUT
 #cmakedefine FLOG_ENABLE_DEBUG
-#cmakedefine FLOG_ENABLE_MPI
 #define FLOG_SERIALIZATION_INTERVAL @FLOG_SERIALIZATION_INTERVAL@
 #endif
 
