### 2025-12-09 5122232
Advertise flog::flush for flog.hh
