### 2025-12-10 f87b29b
Make window size link more precise
