### 2025-12-10 c77dcc3
Merge (some of) profiling.rst into performance.rst
