### 2025-12-10 e97a76c
Remove () from function names
