### 2025-12-10 609a817
Document and deprecate launch_type_t
Add compatibility alias future_kind
Remove unnecessary explicit default arguments
