### 2025-12-13 c28618a
Rewrite parameter checking
Recurse into tuples and vectors
Absorb bad_accessor
Trap resulting impossible conversions for MPI tasks
Remove decay_tuple
Extend tests and comments
