### 2025-12-16 flecsi/flecsi#5 @lindsayad
Thanks @rbberger. I was able to build and install the library fine now
