### 2025-12-16 flecsi/flecsi#6 @lindsayad
Outputting parallelism information
I've enabled FLOG. I'm playing with the poisson example. Is there a way to output parallelism information, e.g. how many processes/threads I'm using?

Refs https://github.com/openjournals/joss-reviews/issues/9333
