### 2026-01-31 flecsi/flecsi#6 @JulienLoiseau
You can use: https://flecsi.github.io/flecsi/api/user/structflecsi_1_1runtime.html
(The API documentation is available on our website flecsi.org.)
From a scheduler, inside an action, you can access the number of ranks you used to start the application like:

```
void
my_action(spec::control_policy & policy) {
  auto & sch = policy.scheduler();
  flog(info) << sch.runtime().process() << "/" << sch.runtime().processes() << '\n';
}
```

Inside a task you can access to which index point and the total number of index points:

```
void
my_task(flecsi::exec::accelerator s,
  /* some fields */) noexcept {
  flog(info) << s.launch().index << "/" << s.launch().size << '\n';
}
```
