### 2026-02-13 faca067
clarify compiler requirements
