### 2026-02-18 bf1dd6f
Add flecsi::comm
Generalize send by using std::invoke
