### 2026-02-20 26d26ac
Add exec::mapping
Split tasks::make out of space_base::bind
Factor out bind_parameters::visit(processor_space_t<Proc>)
