### 2026-02-24 594a018
Add exec::group
Request rank-matching and/or concurrent execution when it is used
Add generic has_param to detect such uses
Add synchronicity parameter to protocol to distinguish MPI tasks
Support rank-matching for all processor types in protocol
Support concurrent tasks in mapper
Trap moving non-portable fields between host and device memory

Discuss what are really effects of rank-matching in Programming Model
Replace "MPI" with "rank-matched" where appropriate in API reference
