### 2026-02-25 00adf06
remove flog stream label
- deprecate add_output_stream overload with stream identifier
- add overload without label
- remove undocumented enable_buffer/disable_buffer
