+++ tutorial/2-control/3-dependencies.cc
@@ -17,6 +17,6 @@ int
 main() {
   const flecsi::run::dependencies_guard dg;
   flecsi::runtime run;
-  flecsi::flog::add_output_stream("clog", std::clog, true);
+  flecsi::flog::add_output_stream(std::clog, true);
   return run.control<control>();
 } // main
