+++ tutorial/2-control/4-state.cc
@@ -66,6 +66,6 @@ int
 main() {
   const flecsi::run::dependencies_guard dg;
   flecsi::runtime run;
-  flecsi::flog::add_output_stream("clog", std::clog, true);
+  flecsi::flog::add_output_stream(std::clog, true);
   return run.control<control>();
 } // main
