+++ flecsi/run/test/cycle.cc
@@ -147,7 +147,7 @@ main(int argc, char ** argv0) {
     return status < flecsi::run::status::clean ? 0 : status;
   }
 
-  flecsi::flog::add_output_stream("clog", std::clog, true);
+  flecsi::flog::add_output_stream(std::clog, true);
 
   status = flecsi::start(control::execute);
 
