### 2026-02-26 #9333 @mprobson
Hello! I am still working my way through it. I've made a few updates and am almost there. I'm finding that I need a concrete deadline to make sure this gets done. Looking ahead, I plan to complete everything by 3/12 at the latest. Thanks for checking back in!

A future note, I saw there was a way to mark 'away' via editorialbot but no way to track my progress up to a point. I think that might be a useful feature to set an internal deadline and get reminders.
