### 2026-02-27 d66a94f
Support std::variant task parameters/arguments
