### 2026-03-02 223c116
Support std::optional task parameters/arguments
