+++ doc/sphinx/news.rst
@@ -37,7 +37,7 @@ New features
 
 * Execution
 
-  * A task parameter can be a ``std::variant`` that includes a FleCSI type such as an accessor.
+  * A task parameter can be a ``std::optional`` of, or a ``std::variant`` that includes, a FleCSI type such as an accessor.
   * ``future_kind`` is an alias for the type of the second ``future`` template parameter.
 
 * Logging
