### 2026-03-04 fd7710a
Store Kokkos execution spaces
Replace execution_space member with kokkos_space trait
Copy them into executor_base, where expose them as kokkos
