### 2026-03-09 542e541
spack: add formatting deps for CI
