### 2026-03-09 4c369c6
spack: add build environment changes for CI
