+++ spack_repo/flecsi/packages/flecsi/package.py
@@ -29,6 +29,20 @@ class Flecsi(Flecsi):
     depends_on("llvm@13 +clang", type="build", when="+format")
     depends_on("git", type="build", when="+format")
 
+    def setup_build_environment(self, env):
+        # build environment changes are needed for CI testing
+        # compiler env vars get put into the generated build-env, which allows
+        # standalones to pick up the right compiler
+        if self.run_tests and self.spec.satisfies("^ucx"):
+            # UCX workaround to avoid misdetecting GPU memory as host memory
+            env.set("UCX_MEMTYPE_CACHE", "n")
+        if self.spec.satisfies("^kokkos +rocm"):
+            env.set("CC", self.spec["hip"].hipcc)
+            env.set("CXX", self.spec["hip"].hipcc)
+        if self.run_tests and self.spec.satisfies("^[virtuals=mpi]openmpi@5:"):
+            # OpenMPI 5.x uses bind-to core by default, limiting us to a single core
+            env.set("OMPI_MCA_hwloc_base_binding_policy", "none")
+
     def cmake_args(self):
         args = super().cmake_args()
         args.append(self.define_from_variant("ENABLE_FORMAT", "format"))
