### 2026-03-09 8d06975
ci: use kessel, require Spack v1.1.x
