### 2026-03-13 #9333 @mprobson
Just a quick update. I was able to install (via Spack) and build (via the GitHub repo) the software. However, I am running into an issue executing the tests on my specific cluster (Unity: https://unity.rc.umass.edu/ ). I'll post the latest error I get below but I will continue working on resolving this test issue.

```sh
spack env create flecsi-dev
spack env activate flecsi-dev
spack install --add --only dependencies flecsi
git clone https://github.com/flecsi/flecsi.git
cd flecsi
mkdir install
mkdir build
cd build
cmake -DFLECSI_BACKEND=mpi -DENABLE_UNIT_TESTS=ON -DCMAKE_INSTALL_PREFIX=../install ..
make -j 4
make install

mrobson_smith_edu@cpu083:~/joss/flecsi/build$ ctest
Test project /home/mrobson_smith_edu/joss/flecsi/build
      Start  1: kernel
 1/18 Test  #1: kernel ...........................***Failed    0.37 sec
Illegal instruction
mrobson_smith_edu@cpu083:~/joss/flecsi/build$ make test
Running tests...
Test project /home/mrobson_smith_edu/joss/flecsi/build
      Start  1: kernel
 1/18 Test  #1: kernel ...........................***Failed    0.37 sec
make: *** [Makefile:71: test] Illegal instruction
```
