### 2026-03-13 #9333 @mprobson
Using `spack load` directly, seems to have solved my issue, i.e. the following works:

```sh
module load mpich/4.2.1
spack load kokkos target=zen2
```

and then continued from there. I'll update my review.
