### 2026-04-09 694d987
ci: re-enable Spack colored output
The new Spack installer make CI output too flat otherwise.
