### 2026-04-09 842569b
ci: explicitly set if tests need to be built
