### 2026-04-20 #9333 @jedbrown
Hi, sorry about my delay (been underwater lately). The paper looks good overall, though I do have some questions before we proceed.

I'm curious what the following could mean because a preconditioner is not an operation that goes *before* a solver, but something that modifies the operator *inside* a solver (e.g., each Krylov iteration of a linear solver, or inside a domain decomposition of a Newton method in ASPIN).

> For example, if an existing application defines a “solver” action, a new developer later can create a “preconditioner” action and insert it before the solver in the DAG without having to modify any other code or interfaces

You could add an "initial guess" action before a solver, but that is not preconditioning, which is more like an extra argument of the solver, thus generating a callback/method invocation from inside the solver.

I think the example invoking "hydrodynamics" and "viscosity" could be confusing since outside of very specific splitting methods, "hydrodynamics" and "viscosity" would not normally be coarse-grained distinct operations (at least not in code intended to have performance via memory locality). Many readers will expect that inviscid and viscous terms are evaluated at quadrature points within the same kernel. If you're referring to particular splitting methods, perhaps that could be clarified or a less-ambiguous example could be used.

"mesh stiffening" seems to be a domain-specific term of art. Perhaps you could pick an example that would be more broadly understood.

"Kernel can execute" (grammar) should say something like "Kernel code can execute" or "Kernels can execute".

It would be informative to explain basics of the programming model by which "no code modifications are needed". Is there a way to express algorithms that use shared memory within this system? Does it always execute as naive parallelism within each kernel invocation? Are there are performance comparisons relative to other projects?

I think the paper should give hints about target granularity sooner in the discussion. For example, projects like MOOSE (Idaho National Lab) have dynamic execution at each quadrature point of finite element integration, but one has to read to the very end to get a sense that this is too fine-grained for FleCSI to be helpful.
