+++ doc/sphinx/news.rst
@@ -46,6 +46,11 @@ Fixed
     (Part of this was a regression in 2.4.1.)
   * CPU tasks properly make use of idle cores.
 
+* MPI backend
+
+  * ``future<void, exec::launch_type_t::index>`` task arguments are accepted.
+    (For portability, note that only recent versions of Legion honor that sort of dependency.)
+
 * HPX backend
 
   * The performance of reduction tasks and certain kinds of single tasks is significantly improved.
