### 2026-04-27 dd22ffc
Use std::span internally
Mark util::span borrowed for compatibility
