+++ flecsi/exec/test/kernel.cc
@@ -12,7 +12,7 @@ const intN::definition<topo::index> array_field;
 
 void
 modify(accelerator s, intN::accessor<wo> a) noexcept {
-  s.executor().forall(i, util::span(*a)) {
+  s.executor().forall(i, std::span(*a)) {
     i = 3;
   };
 }
@@ -20,7 +20,7 @@ modify(accelerator s, intN::accessor<wo> a) noexcept {
 int
 check(intN::accessor<ro> a) noexcept {
   UNIT() {
-    for(auto i : util::span(*a)) {
+    for(auto i : std::span(*a)) {
       EXPECT_EQ(i, 3);
     }
   };
@@ -28,7 +28,7 @@ check(intN::accessor<ro> a) noexcept {
 
 void
 modify_policy(accelerator s, intN::accessor<wo> a) noexcept {
-  s.executor().threads<64, 1>().forall(i, util::span(*a)) {
+  s.executor().threads<64, 1>().forall(i, std::span(*a)) {
     i = 3;
   };
 }
@@ -36,7 +36,7 @@ modify_policy(accelerator s, intN::accessor<wo> a) noexcept {
 int
 check_policy(intN::accessor<ro> a) noexcept {
   UNIT() {
-    for(auto i : util::span(*a)) {
+    for(auto i : std::span(*a)) {
       EXPECT_EQ(i, 3);
     }
   };
@@ -55,7 +55,7 @@ int
 reduce_vec(accelerator s, intN::accessor<ro> a) noexcept {
   UNIT() {
     size_t res =
-      s.executor().reduceall(i, up, util::span(*a), exec::fold::sum, size_t) {
+      s.executor().reduceall(i, up, std::span(*a), exec::fold::sum, size_t) {
       up(i);
     };
     EXPECT_EQ(res, 3 * a.get().size());
@@ -69,7 +69,7 @@ reduce_vec(accelerator s, intN::accessor<ro> a) noexcept {
 
 void
 mdrange_init(accelerator s, intN::accessor<wo> a) noexcept {
-  auto ar = util::span(*a);
+  auto ar = std::span(*a);
   util::mdspan<std::size_t, 2> md_ar(ar.data(), {5, 2});
   s.executor().forall(mi, (mdiota_view(md_ar, full_range(), prefix_range{2}))) {
     auto [i, j] = mi;
@@ -80,7 +80,7 @@ mdrange_init(accelerator s, intN::accessor<wo> a) noexcept {
 int
 check_mdrange(intN::accessor<ro> a) noexcept {
   UNIT() {
-    for(auto i : util::span(*a)) {
+    for(auto i : std::span(*a)) {
       EXPECT_EQ(i, 3);
     }
   };
@@ -89,7 +89,7 @@ check_mdrange(intN::accessor<ro> a) noexcept {
 int
 reduce_mdrange_vec(accelerator s, intN::accessor<rw> a) noexcept {
   UNIT() {
-    auto ar = util::span(*a);
+    auto ar = std::span(*a);
     util::mdspan<std::size_t, 2> md_ar(ar.data(), {5, 2});
     size_t res = s.executor().reduceall(mi,
       up,
