### 2026-04-29 e3f5c02
Populate scalar_access on demand
